Insere validação para marca

+++ app/Http/Controllers/MarcaController.php
@@ -20,7 +20,7 @@ class MarcaController extends Controller
     {
         // $marcas = Marca::all();
         $marcas  =$this->marca->all();
-        return $marcas;
+        return response()->json($marcas, 200);
     }
 
     /**
@@ -42,8 +42,19 @@ class MarcaController extends Controller
     public function store(Request $request)
     {
         // $marca = Marca::create($request->all());
-        $marca = $this->marca->create($request->all());
-        return $marca;
+        // nome
+        // marca
+
+        $request->validate($this->marca->rules(), $this->marca->feedback());
+        // stateless
+
+        // dd($request->nome);
+        // dd($request->get('nome'));
+        // dd($request->imagem);
+        dd($request->file('imagem'));
+
+        // $marca = $this->marca->create($request->all());
+        return response()->json($marca, 201);
     }
 
     /**
@@ -55,7 +66,11 @@ class MarcaController extends Controller
     public function show($id)
     {
         $marca = $this->marca->find($id);
-        return $marca;
+        if($marca === null) {
+            return response()->json(['erro' => 'Recurso pesquisado não existe'], 404); 
+        }
+
+        return response()->json($marca, 200);
     }
 
     /**
@@ -78,12 +93,30 @@ class MarcaController extends Controller
      */
     public function update(Request $request, $id)
     {
-        // print_r($request->all()); 
-        // print_r($marca->getAttributes()); 
-        // $marca->update($request->all());
         $marca = $this->marca->find($id);
+
+        if($marca === null) {
+            return response()->json(['erro' => 'Impossível realizar a atualização. O recurso solicitado não existe'], 404);
+        }
+
+        if($request->method() === 'PATCH') {
+            // return ['teste' => 'Verbo PATCH'];
+
+            $regrasDinamicas = array();
+
+            foreach($marca->rules() as $input => $regra) {
+                if(array_key_exists($input, $request->all())){
+                    $regrasDinamicas[$input] = $regra;
+                }
+            }
+
+            $request->validate($regrasDinamicas, $marca->feedback());
+        } else {
+            $request->validate($marca->rules(), $marca->feedback());
+        }
+
         $marca->update($request->all());
-        return $marca;
+        return response()->json($marca, 200);
     }
 
     /**
@@ -95,7 +128,12 @@ class MarcaController extends Controller
     public function destroy($id)
     {
         $marca = $this->marca->find($id);
+
+        if($marca === null) {
+            return response()->json(['erro' => 'Impossível realizar a exclusão. O recurso solicitado não existe'], 404);
+        }
+        
         $marca->delete();
-        return ['msg' => 'A marca foi removida com sucesso!'];
+        return response()->json(['msg' => 'A marca foi removida com sucesso!'], 200);
     }
 }
